requirements.txt
streamlit>=1.29.0
google-adk==1.12.0
python-dotenv==1.0.0
pydantic>=2.0.0,<3.0.0
nest_asyncio>=1.5.0
litellm>=1.17.9
